requirements.txt
absl-py==0.6.1
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.9.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.0
boto==2.49.0
boto3==1.9.82
botocore==1.12.82
Bottleneck==1.2.1
bz2file==0.98
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.5.5
cryptography==2.5
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
dask==2.1.0
decorator==4.3.2
defusedxml==0.6.0
distributed==2.1.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
Flask==1.1.1
gast==0.2.0
gensim==3.7.1
gevent==1.4.0
greenlet==0.4.15
grpcio==1.16.1
h5py==2.9.0
heapdict==1.0.0
html2text==2018.1.9
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.17
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
json5==0.8.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
kaggle==1.5.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
llvmlite==0.29.0
locket==0.2.0
lxml==4.3.4
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.3
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.45.0
numexpr==2.6.9
numpy==1.15.4
numpydoc==0.9.1
olefile==0.46
openpyxl==2.6.2
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.2
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patool==1.12
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.12.0
ply==3.11
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
psutil==5.6.3
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.7.5
python-slugify==3.0.2
pytz==2018.9
PyWavelets==1.0.3
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.7
qtconsole==4.5.2
QtPy==1.8.0
requests==2.21.0
rope==0.14.0
ruamel-yaml==0.15.46
s3transfer==0.1.13
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.0
snowballstemmer==1.9.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.5
statsmodels==0.10.0
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.10.0
torch==1.0.1
torchvision==0.2.2
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
typed-ast==1.3.4
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwings==0.15.8
xlwt==1.3.0
zict==1.0.0
zipp==0.5.1
